refactor: moved multiobject to output file.
